+++ spec/fixtures/high_priority_template.rb
@@ -3,4 +3,6 @@ Racker::Processor.register_template do |t|
     :iso_url => 'priority.img',
     :password => '~~',
   }
+  t.description = 'some description'
+  t.min_packer_version = '1.1.1'
 end